Add 'evaluate' function for multiple metrics
